refactor: Move compoments/header to features/header dir
