+++ lib/app/home_page.dart
@@ -1,10 +1,10 @@
 import 'dart:math';
 
-import 'package:confwebsite2023/core/components/header/header_item_button_data.dart';
-import 'package:confwebsite2023/core/components/header/header_widget.dart';
 import 'package:confwebsite2023/core/components/responsive_widget.dart';
 import 'package:confwebsite2023/core/theme/baseline_color_scheme.dart';
 import 'package:confwebsite2023/features/footer/ui/footer.dart';
+import 'package:confwebsite2023/features/header/data/header_item_button_data.dart';
+import 'package:confwebsite2023/features/header/ui/header_widget.dart';
 import 'package:confwebsite2023/features/hero_section/ui/hero_section.dart';
 import 'package:confwebsite2023/features/news/ui/news_section.dart';
 import 'package:confwebsite2023/features/staff/ui/staff_section.dart';
